lab4src/requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
tensorflow==2.18.0
tf-keras==2.18.0
sentencepiece==0.2.0
transformers==4.46.2
scipy==1.14.1
tiktoken==0.8.0
pandas==2.2.3
datasets==3.1.0
accelerate==1.1.1

python-dotenv==1.0.1
aiogram==3.14.0

pyyaml==6.0.2
pathlib==1.0.1

bitsandbytes==0.42.0
einops==0.8.0
gpt4all==2.8.2

googletrans==4.0.0-rc1